Add skeleton of satellite server App Engine app.
This change includes an unimplemented HTTP server that uses basic http auth for
authentication and authorization.

The Authenticator interface can be used to switch between basic auth and
any other auth methods we want to support in the future. The settings
should be configurable by app admins, and this configuration will be
stored in the App Engine datastore.